Support method deletion in `static_methods()` invalidation.
Add all matching method instances as backedges to `static_methods()`, so
that deleting a method (which apparently only invalidates functions that
inlined that method), will still invalidate our `static_methods()`
function.

Adds test for method deletion. :)

Note that there is a problem with detecting covering methods when the
provided type includes Varargs, so we currently oversubscribe to more
MethodInstances than we need to. :/