Support method deletion in `static_methods()` invalidation. (#34)
* Support method deletion in `static_methods()` invalidation.

Add all matching method instances as backedges to `static_methods()`, so
that deleting a method (which apparently only invalidates functions that
inlined that method), will still invalidate our `static_methods()`
function.

Adds test for method deletion. :)

Note that there is a problem with detecting covering methods when the
provided type includes Varargs, so we currently oversubscribe to more
MethodInstances than we need to. :/

* Support method deletions in static_method_count() as well

* Add unit test for static_methods with type signatures

Including tests for Varargs in the type

Co-authored-by: Lyndon White <oxinabox@ucc.asn.au>

* Bump version number to v0.1.5

Co-authored-by: Lyndon White <oxinabox@ucc.asn.au>